test: Combine test cases for debug-ability
In ruby I wouldn't mind having exceptions. However in Go I find that a
panic causes the output of your tests to be pretty hard to read. I had
this happen when I had a separate test that showed there were 2 files on
disk, and that those files both had content. When there were 0 files
detected, the latter had a panic which was hard to debug.

To make this easier to manage, these test cases have been merged
together. This should make the test suite run a bit faster given we're
asserting a few things about the operation of the code under test.